Moved extraction and injection classes in their respective namespace.
