+++ ExtractionStrategy/ExtractionStrategyInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace Innmind\Reflection;
+namespace Innmind\Reflection\ExtractionStrategy;
 
 interface ExtractionStrategyInterface
 {
