+++ InjectionStrategy/ReflectionStrategy.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Reflection\InjectionStrategy;
 
-use Innmind\Reflection\InjectionStrategyInterface;
 use Innmind\Reflection\Exception\LogicException;
 
 class ReflectionStrategy implements InjectionStrategyInterface
