+++ Tests/InjectionStrategy/ReflectionStrategyTest.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Reflection\Tests\InjectionStrategy;
 
-use Innmind\Reflection\InjectionStrategyInterface;
 use Innmind\Reflection\InjectionStrategy\ReflectionStrategy;
 
 class ReflectionStrategyTest extends \PHPUnit_Framework_TestCase
