+++ Tests/ReflectionClassTest.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\Reflection\Tests;
 
 use Innmind\Reflection\ReflectionClass;
-use Innmind\Reflection\InjectionStrategyInterface;
+use Innmind\Reflection\InjectionStrategy\InjectionStrategyInterface;
 use Innmind\Reflection\InjectionStrategy\SetterStrategy;
 use Innmind\Reflection\InjectionStrategy\NamedMethodStrategy;
 use Innmind\Reflection\InjectionStrategy\ReflectionStrategy;
